WIP: Clean code and add more error check
